refactor: 💡 Rename and moved some files
✅ Closes: https://hashicorp.atlassian.net/browse/ICU-17409
